Fix: breakpoint numbers were confused with non-source code By: Hiro Sugawara <hiro@lynx.com>
The bug was: In the following simple program,
from the GDB pane/window, type "break exit" or "break printf" (or "break
any_non_source_code"), to set breakpoint #1;
then set a breakpoint with the mouse at the "printf" statement to set
breakpoint #2.
Move the mouse over to the printf statement; "Breakpoint #1" pops up
whereas "#2" is expected.
Deleting/disabling the breakpoint on the printf statement actually
affects breakpoint #1 ("exit" or "printf" in this case).

#include <stdio.h>
main()
{
  printf("hello world\n");
}


git-svn-id: svn://svn.savannah.gnu.org/ddd/trunk@5887 1d73a310-8f1f-0410-b01a-dab0232e19a9